+++ scripts/discord.sh
@@ -1,7 +1,9 @@
 #!/bin/bash
 set -xe
 
-rm /var/discord.deb -f
-wget -O /tmp/discord.deb "https://discordapp.com/api/download?platform=linux&format=deb"
-sudo apt-get install -y discord
+if ! type discord > /dev/null 2>&1; then
+    rm /var/discord.deb -f
+    wget -O /tmp/discord.deb "https://discordapp.com/api/download?platform=linux&format=deb"
+    sudo apt-get install -y /tmp/discord.deb
+fi
